merged: changes from trunk
notes:

re-generated English strings.po from strings.xml and merged in trunk language string changes
fixed up skin variables patches
removed custom has addon code
fixed up language loading

may be some new bugs.

+++ branches/addons/xbmc/FileSystem/CircularCache.cpp
@@ -1,6 +1,6 @@
 /*
  *      Copyright (C) 2005-2013 Team XBMC
- *      http://www.xbmc.org
+ *      http://xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -13,9 +13,8 @@
  *  GNU General Public License for more details.
  *
  *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, write to
- *  the Free Software Foundation, 675 Mass Ave, Cambridge, MA 02139, USA.
- *  http://www.gnu.org/copyleft/gpl.html
+ *  along with XBMC; see the file COPYING.  If not, see
+ *  <http://www.gnu.org/licenses/>.
  *
  */
 
@@ -106,7 +105,7 @@ int CCircularCache::WriteToCache(const char *buf, size_t len)
   m_end += len;
 
   // drop history that was overwritten
-  if(m_end - m_beg > m_size)
+  if(m_end - m_beg > (int64_t)m_size)
     m_beg = m_end - m_size;
 
   m_written.Set();
